# HTTPServer/http-api-spec.yaml
openapi: 3.0.0
info:
  title: "Game Logic Gateway API"
  description: "클라이언트와 HTTPS 서버를 위한 REST API 명세서"
  version: "1.0.0"

servers:
  - url: http://localhost:3000
    description: "로컬 개발 환경"

paths:
  # 버전 및 상태 확인 API
  /api/version:
    get:
      summary: "클라이언트 버전 및 점검 상태 확인"
      tags:
        - System
      description: "게임 실행 시 클라이언트 버전이 유효한지, 서버가 점검 중은 아닌지 확인합니다."
      responses:
        '200':
          description: "조회 성공"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/VersionData'

  # 계정 생성 API (회원가입)
  /api/signup:
    post:
      summary: "신규 계정 생성"
      tags:
        - Auth
      description: "ID와 비밀번호를 받아 새로운 유저 계정을 생성합니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: "계정 생성 성공 (자동 로그인 처리되어 세션 키 발급)"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/AuthData'
        '409':
          description: "이미 존재하는 ID (Conflict)"

  # 로그인 API
  /api/login:
    post:
      summary: "로그인 요청"
      tags:
        - Auth
      description: "ID와 비밀번호로 로그인을 시도합니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: "로그인 성공"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/AuthData'
        '401':
          description: "인증 실패 (비밀번호 불일치 또는 존재하지 않는 ID)"

  # 매치메이킹 시작
  /api/game/match/start:
    post:
      summary: "매치메이킹 시작 요청"
      tags:
        - Game
      security:
        - SessionAuth: [] 
      description: "사용할 맵과 아이템을 서버에 전달하고, 매치메이킹 대기열에 진입(티켓 발급)합니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameReadyRequest'
      responses:
        '200':
          description: "대기열 진입 성공 (매칭 확인용 티켓 ID 반환)"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/MatchTicketData'

  # 매치메이킹 상태 확인 (Polling)
  /api/game/match/status:
    get:
      summary: "매치메이킹 상태 확인 (폴링)"
      tags:
        - Game
      security:
        - SessionAuth: []
      description: "발급받은 티켓 ID를 이용해 주기적으로 매칭이 성사되었는지 확인합니다."
      parameters:
        - name: ticketId
          in: query
          required: true
          schema:
            type: string
          description: "매치메이킹 시작 시 발급받은 티켓 ID"
      responses:
        '200':
          description: "상태 조회 성공. 매칭 완료 시 접속 정보가 함께 반환됩니다."
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/MatchStatusData'

  # 매치메이킹 취소
  /api/game/match/cancel:
    post:
      summary: "매치메이킹 취소 요청"
      tags:
        - Game
      security:
        - SessionAuth: []
      description: "대기 중인 매치메이킹을 취소합니다."
      requestBody: # [수정] Query Parameter에서 Body로 변경
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ticketId
              properties:
                ticketId:
                  type: string
                  description: "취소할 매칭 티켓 ID"
                  example: "ticket_1234abcd"
      responses:
        '200':
          description: "취소 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
        '409':
          description: "이미 매칭이 성사되어 취소 불가능"

components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Id
      description: "로그인 시 발급받은 세션 키를 입력해주세요."

  schemas:
    # ----------------------------------------------------
    # 공통 응답 (Base Wrapper)
    # ----------------------------------------------------
    CommonResponse:
      type: object
      required: [success, code]
      properties:
        success:
          type: boolean
          example: true
        code:
          type: integer
          example: 200
        data:
          type: object
          nullable: true
          description: "API 호출 결과 데이터 (API마다 구조가 다름)"
        error:
          $ref: '#/components/schemas/ErrorDetail'
          
    ErrorDetail:
      type: object
      nullable: true
      properties:
        message: { type: string, example: "Invalid request" }
        code: { type: string, example: "ERR_BAD_REQUEST" }

    # ----------------------------------------------------
    # 공통 응답에 담을 진짜 내용물
    # ----------------------------------------------------
    VersionData:
      type: object
      properties:
        latestVersion:
          type: string
          example: "1.0.5"
        isMaintenance:
          type: boolean
          description: "서버 점검 중 여부"
          example: false

    AuthData:
      type: object
      properties:
        sessionId:
          type: string
          description: "API 인증에 사용할 세션 키"
          example: "abc123xyz456"

    # ----------------------------------------------------
    # 요청 본문 스키마들 (Request Bodies)
    # ----------------------------------------------------
    AuthRequest:
      type: object
      required:
        - id
        - password
      properties:
        id:
          type: string
          description: "아이디. alphanumeric 4자 이상, 16자 이하"
          example: "tetepiti149"
          minLength: 4
          maxLength: 16
          pattern: "^[a-zA-Z0-9]+$"
        password:
          type: string
          format: password
          description: "유저 비밀번호. 서버에서 해싱 처리 예정."
          example: "q1w2e3r4!@"
          minLength: 4
          maxLength: 16

    GameReadyRequest:
      type: object
      required:
        - mapId
        - equippedItems
      properties:
        mapId:
          type: string
          description: "진입할 맵의 식별자"
          example: "0"
        equippedItems: # [수정] items -> equippedItems 로 변경
          type: array
          description: "인게임에서 사용할 인벤토리 아이템 목록 (ID와 수량)"
          items:
            type: object
            required:
              - itemId
              - quantity
            properties:
              itemId:
                type: integer # [수정] string -> integer
                description: "아이템 식별자 ID"
                example: 101
              quantity:
                type: integer
                description: "가져갈 아이템의 개수"
                example: 5
          example: 
            - itemId: 101
              quantity: 5
            - itemId: 105
              quantity: 2

    MatchTicketData:
      type: object
      properties:
        ticketId:
          type: string
          description: "상태 조회에 사용할 매칭 대기표 ID"
          example: "ticket_987654321"

    MatchStatusData:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: "현재 매칭 상태 (WAITING: 대기중, SUCCESS: 매칭 성공, FAILED: 실패)"
          enum: [WAITING, SUCCESS, FAILED]
          example: "SUCCESS"
        ip: # [수정] udpServerIp -> ip
          type: string
          description: "접속해야 할 인게임 UDP 서버 IP (상태가 SUCCESS일 때만 존재)"
          example: "127.0.0.1"
        port: # [수정] udpServerPort -> port
          type: integer
          description: "접속해야 할 인게임 UDP 서버 포트"
          example: 7777
        roomToken:
          type: string
          description: "인게임 진입 시 사용할 방 입장 권한 토큰"
          example: "room_token_abc"